Adding pundit and authorization for updating/deleting posts and comments

+++ app/controllers/posts_controller.rb
@@ -25,6 +25,7 @@ class PostsController < ApplicationController
 
   def update
     post = Post.find(params[:id])
+    authorize post
 
     if post.update(post_params)
       render json: post
@@ -35,6 +36,8 @@ class PostsController < ApplicationController
 
   def destroy
     post = Post.find(params[:id])
+    authorize post
+
     post.destroy
 
     head :no_content
